Fix error when loading Qml imports from a remote Url
Currently qmldir files may be parsed twice when retrieved via remote
urls, once for the full url, and once for '.', resulting in an
error. While loading twice is not ideal, QQmlDirParser seems to
support it, as all its other members are cleared at the beginning
of parse(). So this change just clears _typeNamespace along with
everything else at the start of parsing.

Task-number: QTBUG-30467

Change-Id: Ic6f199b914f45dfc47f41a44d5e1a8a2767e68f2